util: rearrange metrics
